add login test
